pyproject.toml
[project]
name = "trading-bot"
version = "0.1.0"
description = "Multi-strategy paper-first trading bot platform with comparative dashboard."
requires-python = ">=3.11"
readme = "README.md"
# Slim base — what the live worker needs PLUS the FastAPI dashboard backend.
dependencies = [
    "alpaca-py>=0.30.0",
    "apscheduler>=3.10.4",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "python-dotenv>=1.0",
    "sqlalchemy>=2.0",
    "pandas>=2.1",
    "numpy>=1.26",
    "yfinance>=0.2.40",
    "requests>=2.31",
    "httpx>=0.27",
    "click>=8.1",
    "rich>=13.7",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
]

[project.optional-dependencies]
# Backtest harness — vectorbt pulls in numba/llvmlite which are heavy; keep them out of prod.
backtest = [
    "vectorbt>=0.26",
    "ta>=0.11",
    "optuna>=3.5",
]
# News/sentiment FinBERT pipeline.
sentiment = [
    "transformers>=4.40",
    "torch>=2.2",
]
# Research agent: scrape AI-trading content from social/web/video, synthesize with Gemini.
# Heavy and optional — installs PydanticAI + Gemini SDK + source clients.
research = [
    "pydantic-ai-slim[google]>=0.0.36",
    "google-genai>=0.3",
    "asyncpraw>=7.7",
    "youtube-transcript-api>=0.6.2",
    "yt-dlp>=2024.4.9",
    "feedparser>=6.0",            # arXiv RSS, generic feeds
    "tavily-python>=0.5",         # general web search (1k free/mo)
    "selectolax>=0.3.21",          # fast HTML → text extraction
    "tenacity>=8.2",              # retry/backoff
    "logfire>=0.30",              # PydanticAI observability (optional)
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "vectorbt>=0.26",
    "ta>=0.11",
    "optuna>=3.5",
]

[project.scripts]
trading-bot = "src.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "dashboard*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"
